#44 refactor: PortfolioHoldingLots 컴포넌트 분리
