+++ fe/src/pages/PortfolioPage.tsx
@@ -1,6 +1,6 @@
 import usePortfolioDetailsQuery from "@api/portfolio/queries/usePortfolioDetailsQuery";
 import plusIcon from "@assets/icons/plus.svg";
-import PortfolioHoldingsTable from "@components/Portfolio/PortfolioHoldingsTable";
+import PortfolioHoldingsTable from "@components/Portfolio/PortfolioHoldings/PortfolioHoldingsTable";
 import { Box, Button, Typography } from "@mui/material";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
